lil refactor 2
